requirements.txt
playwright
requests
python-dotenv
agent-framework
pydantic
colorama